feature(services): se agregaron metodos en los services, se corrigieron endpoints

+++ src/main/java/com/itm/ecosurprise/dtos/ProductoDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ProductoDTO {
     private int id;
     private String nombre;
-    private double precio;
+    private int precio;
 }
